pyproject.toml
[project]
name = "aimq"
version = "0.1.2"
description = "A robust message queue processor for Supabase pgmq with AI-powered document processing capabilities"
authors = [{name = "AIMQ Contributors"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<=3.13"
keywords = [
    "ai",
    "ocr",
    "document-processing",
    "supabase",
    "pgmq",
    "queue",
    "machine-learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "easyocr>=1.7.1,<2.0.0",
    "torch>=2.9.0,<3.0.0", # Exclude vulnerable 2.8.0 (CVE-2025-55551)
    "python-dotenv>=1.0.0,<2.0.0",
    "supabase>=2.10.0,<3.0.0",
    "typer>=0.9.0,<1.0.0",
    "rich>=13.7.0,<15.0.0",
    "pillow>=10.2.0,<13.0.0",
    "openai>=1.57.1,<2.0.0",
    "langchain>=0.3.10,<0.4.0",
    "langchain-openai>=0.2.12,<0.3.0",
    "langchain-text-splitters>=0.3.9,<1.0.0",
    "pydantic-settings>=2.6.1,<3.0.0",
    "filetype>=1.2.0,<2.0.0",
    "humanize>=4.11.0,<5.0.0",
    "pypdfium2>=4.30.0,<5.0.0",
    "mkdocs-dracula-theme>=1.0.7,<2.0.0",
    "tomlkit>=0.13.2,<1.0.0",
    "mistralai>=1.9.10,<2.0.0",
    "langchain-mistralai>=0.2.11,<0.3.0",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langchain-core>=0.3.79",
    "psycopg-binary>=3.2.12",
    "python-magic>=0.4.27,<1.0.0",
    "tenacity>=8.2.0,<10.0.0",
    "resend>=2.19.0",
]


[project.scripts]
aimq = "aimq.commands:app"

[project.urls]
Homepage = "https://github.com/bldxio/aimq"
Repository = "https://github.com/bldxio/aimq"
Documentation = "https://bldxio.github.io/aimq"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aimq"]
artifacts = [
    "src/aimq/commands/shared/templates/**/*",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --timeout=30 --timeout-method=thread"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
# Default timeout for all tests (can be overridden per test)
timeout = 30
timeout_method = "thread"
filterwarnings = [
    # Ignore Mistral model parameter warnings (expected when using mistral_model)
    "ignore:WARNING! mistral_model is not default parameter:UserWarning",
    # Ignore psycopg_pool deprecation warning (will be fixed in future psycopg release)
    "ignore:the default for the ConnectionPool 'open' parameter:DeprecationWarning:psycopg_pool.pool",
]

[tool.coverage.run]
branch = true
source = ["aimq"]
omit = ["src/aimq/commands/shared/templates/*"]
parallel = true
concurrency = ["thread", "multiprocessing"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
show_missing = true
skip_empty = true

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.5,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "ipykernel>=6.29.5,<8.0.0",
    "mkdocs>=1.6.1,<2.0.0",
    "mkdocs-material>=9.5.49,<10.0.0",
    "mkdocstrings[python]>=0.27.0,<1.0.0",
    "mkdocs-gen-files>=0.5.0,<1.0.0",
    "freezegun>=1.5.1,<2.0.0",
    "tox>=4.24.1,<5.0.0",
    "black>=25.9.0,<26.0.0",
    "isort>=5.13.0,<6.0.0",
    "flake8>=7.0.0,<8.0.0",
    "mypy>=1.8.0,<2.0.0",
    "pre-commit>=4.3.0",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.8.0",
]
